+++ CHANGELOG.md
@@ -38,8 +38,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   (<https://github.com/openvinotoolkit/datumaro/pull/684>,
    <https://github.com/openvinotoolkit/datumaro/pull/686>,
    <https://github.com/openvinotoolkit/datumaro/pull/687>)
-- YOLO format now supports almost any subset names, except of
-  just `train` and `valid`
+- YOLO format now supports almost any subset names, except `backup`, `names` and `classes`
+  (instead of just `train` and `valid`). The reserved names now raise an error on exporting.
   (<https://github.com/openvinotoolkit/datumaro/pull/688>)
 
 ### Deprecated